fix : previous error and reflect Input component 기존의 에러를 해결하고, 일반 input을 Input컴포넌트로 교체하였습니다. 추가적으로 login 을 signIn register는 signUp으로 교체하였습니다.
