+++ src/components/atoms/input/index.tsx
@@ -3,7 +3,7 @@ import { forwardRef } from 'react';
 import { ErrorMessage } from './styles';
 
 interface InputProps extends React.DetailedHTMLProps<React.InputHTMLAttributes<HTMLInputElement>, HTMLInputElement> {
-  error: string | undefined;
+  error?: string;
 }
 
 const Input = forwardRef<HTMLInputElement, InputProps>(({ error, ...props }, ref) => {
