+++ src/components/templates/signIn/index.type.ts
@@ -1,4 +1,4 @@
-export interface LoginFormInterface {
+export interface SignInFormInterface {
   email: string;
   password: string;
 }
@@ -12,6 +12,15 @@ export interface InvalidFormInterface {
   email?: InvalidType | undefined;
   password?: InvalidType | undefined;
 }
+
+export const SignInData = {
+  email: {
+    required: '이메일을 입력해주세요.',
+  },
+  password: {
+    required: '비밀번호를 입력해주세요.',
+  },
+};
 
 export const SignInType = {
   email: { placeholder: '이메일', type: 'text' },
