+++ src/pages/_app.tsx
@@ -1,8 +1,8 @@
 import '@styles/globals.css';
 import type { AppProps } from 'next/app';
+import { RecoilRoot } from 'recoil';
 
 import Layout from '@components/Layout';
-import { RecoilRoot } from 'recoil';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
