--- src/pages/login/index.tsx
@@ -1,7 +1,0 @@
-import LoginTemp from '@components/templates/login';
-
-const Login = () => {
-  return <LoginTemp />;
-};
-
-export default Login;
