--- src/pages/register/index.tsx
@@ -1,6 +1,0 @@
-import RegisterTemp from '@templates/register/index';
-const Register = () => {
-  return <RegisterTemp />;
-};
-
-export default Register;
