+++ src/pages/signIn/index.tsx
@@ -1,0 +1,7 @@
+import SignInTemp from '@components/templates/signIn';
+
+const SignIn = () => {
+  return <SignInTemp />;
+};
+
+export default SignIn;
